android_app/requirements-mobile.txt
kivy==2.3.0
requests==2.32.3
python-socketio==5.11.4
